Used short notation for uint32_t and others.

+++ core/enums/direction.h
@@ -1,10 +1,10 @@
 #pragma once
 
-#include <cstdint>
+#include "core/consts.h"
 
 namespace Enum {
 
-enum class Direction : uint8_t {
+enum class Direction : u32 {
 	Reversed,
 	Normal
 };
